DigitRecognition/requirements.txt
tensorflow>=2.0.0
numpy>=1.18.0
matplotlib>=3.1.0
scikit-learn>=0.22.0
pillow>=7.0.0
opencv-python>=4.5.0